support 3d bc model training

+++ empanada/inference/watershed.py
@@ -2,7 +2,7 @@
 Copied and modified from: https://github.com/zudi-lin/pytorch_connectomics/blob/b6457ea4bc7d9b01ef3a00781dff252ab5d4a1d3/connectomics/utils/process.py
 """
 
-
+import sys
 import heapq
 import numba
 import numpy as np
@@ -156,7 +156,7 @@ def bc_watershed(
     foreground = (semantic > int(255*thres3))
     
     # prepare seeds
-    seed = connected_components(seed_map)
+    seed = connected_components(seed_map.astype(np.uint8))
     seed = size_threshold(seed, seed_thres)
     
     if use_mask_wts:
